Refactor data source for edge configuration
