--- castai/data_source_edge_configurations.go
@@ -1,377 +1,0 @@
-package castai
-
-import (
-	"context"
-	"fmt"
-	"net/http"
-
-	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
-	"github.com/hashicorp/terraform-plugin-framework/attr"
-	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
-	"github.com/hashicorp/terraform-plugin-framework/types"
-
-	"github.com/castai/terraform-provider-castai/castai/sdk/omni"
-)
-
-var (
-	_ datasource.DataSource              = (*edgeConfigurationsDataSource)(nil)
-	_ datasource.DataSourceWithConfigure = (*edgeConfigurationsDataSource)(nil)
-)
-
-type edgeConfigurationsDataSource struct {
-	client *ProviderConfig
-}
-
-type edgeConfigurationsDataSourceModel struct {
-	ID             types.String                 `tfsdk:"id"`
-	OrganizationID types.String                 `tfsdk:"organization_id"`
-	ClusterID      types.String                 `tfsdk:"cluster_id"`
-	EdgeLocationID types.String                 `tfsdk:"edge_location_id"`
-	Name           types.String                 `tfsdk:"name"`
-	Configurations []edgeConfigurationDataModel `tfsdk:"configurations"`
-}
-
-type edgeConfigurationDataModel struct {
-	ID             types.String               `tfsdk:"id"`
-	Name           types.String               `tfsdk:"name"`
-	EdgeLocationID types.String               `tfsdk:"edge_location_id"`
-	Default        types.Bool                 `tfsdk:"default"`
-	UserDataBase64 types.String               `tfsdk:"user_data_base64"`
-	Aws            *awsConfigurationDataModel `tfsdk:"aws"`
-	Gcp            *gcpConfigurationDataModel `tfsdk:"gcp"`
-	Oci            *ociConfigurationDataModel `tfsdk:"oci"`
-	Custom         types.Map                  `tfsdk:"custom"`
-}
-
-type awsConfigurationDataModel struct {
-	ImageID         types.String `tfsdk:"image_id"`
-	BootDiskSizeGiB types.Int64  `tfsdk:"boot_disk_size_gib"`
-	Tags            types.Map    `tfsdk:"tags"`
-}
-
-type gcpConfigurationDataModel struct {
-	ImageID         types.String `tfsdk:"image_id"`
-	BootDiskSizeGiB types.Int64  `tfsdk:"boot_disk_size_gib"`
-	Labels          types.Map    `tfsdk:"labels"`
-}
-
-type ociConfigurationDataModel struct {
-	ImageID         types.String `tfsdk:"image_id"`
-	BootDiskSizeGiB types.Int64  `tfsdk:"boot_disk_size_gib"`
-	Tags            types.Map    `tfsdk:"tags"`
-}
-
-func newEdgeConfigurationsDataSource() datasource.DataSource {
-	return &edgeConfigurationsDataSource{}
-}
-
-func (d *edgeConfigurationsDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_edge_configurations"
-}
-
-func (d *edgeConfigurationsDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, resp *datasource.SchemaResponse) {
-	resp.Schema = schema.Schema{
-		Description: "Retrieve information about CAST AI edge configurations for a specific edge location",
-		Attributes: map[string]schema.Attribute{
-			"id": schema.StringAttribute{
-				Optional:    true,
-				Computed:    true,
-				Description: "Filter configurations by ID",
-				Validators: []validator.String{
-					stringvalidator.LengthAtLeast(1),
-				},
-			},
-			"organization_id": schema.StringAttribute{
-				Required:    true,
-				Description: "CAST AI organization ID",
-			},
-			"cluster_id": schema.StringAttribute{
-				Required:    true,
-				Description: "CAST AI cluster ID",
-			},
-			"edge_location_id": schema.StringAttribute{
-				Required:    true,
-				Description: "The ID of the edge location",
-			},
-			"name": schema.StringAttribute{
-				Optional:    true,
-				Description: "Filter configurations by name",
-				Validators: []validator.String{
-					stringvalidator.LengthAtLeast(1),
-				},
-			},
-			"configurations": schema.ListNestedAttribute{
-				Computed:    true,
-				Description: "List of edge configurations",
-				NestedObject: schema.NestedAttributeObject{
-					Attributes: map[string]schema.Attribute{
-						"id": schema.StringAttribute{
-							Computed:    true,
-							Description: "The ID of the edge configuration",
-						},
-						"name": schema.StringAttribute{
-							Computed:    true,
-							Description: "The name of the edge configuration",
-						},
-						"edge_location_id": schema.StringAttribute{
-							Computed:    true,
-							Description: "The ID of the edge location the configuration belongs to",
-						},
-						"default": schema.BoolAttribute{
-							Computed:    true,
-							Description: "Whether this is the default configuration",
-						},
-						"user_data_base64": schema.StringAttribute{
-							Computed:    true,
-							Description: "Base64 encoded user data for edge bootstrap",
-						},
-						"aws": schema.SingleNestedAttribute{
-							Computed:    true,
-							Description: "AWS specific configuration",
-							Attributes: map[string]schema.Attribute{
-								"image_id": schema.StringAttribute{
-									Computed:    true,
-									Description: "AWS AMI ID or name filter for edge creation",
-								},
-								"boot_disk_size_gib": schema.Int64Attribute{
-									Computed:    true,
-									Description: "Boot disk size in GiB",
-								},
-								"tags": schema.MapAttribute{
-									Computed:    true,
-									Description: "Instance/VM tags",
-									ElementType: types.StringType,
-								},
-							},
-						},
-						"gcp": schema.SingleNestedAttribute{
-							Computed:    true,
-							Description: "GCP specific configuration",
-							Attributes: map[string]schema.Attribute{
-								"image_id": schema.StringAttribute{
-									Computed:    true,
-									Description: "GCP image ID or family for edge creation",
-								},
-								"boot_disk_size_gib": schema.Int64Attribute{
-									Computed:    true,
-									Description: "Boot disk size in GiB",
-								},
-								"labels": schema.MapAttribute{
-									Computed:    true,
-									Description: "Instance/VM labels",
-									ElementType: types.StringType,
-								},
-							},
-						},
-						"oci": schema.SingleNestedAttribute{
-							Computed:    true,
-							Description: "OCI specific configuration",
-							Attributes: map[string]schema.Attribute{
-								"image_id": schema.StringAttribute{
-									Computed:    true,
-									Description: "OCI image ID or name filter for edge creation",
-								},
-								"boot_disk_size_gib": schema.Int64Attribute{
-									Computed:    true,
-									Description: "Boot disk size in GiB",
-								},
-								"tags": schema.MapAttribute{
-									Computed:    true,
-									Description: "Instance/VM tags",
-									ElementType: types.StringType,
-								},
-							},
-						},
-						"custom": schema.MapAttribute{
-							Computed:    true,
-							Description: "Custom cloud specific configuration tags",
-							ElementType: types.StringType,
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func (d *edgeConfigurationsDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, resp *datasource.ConfigureResponse) {
-	if req.ProviderData == nil {
-		return
-	}
-
-	client, ok := req.ProviderData.(*ProviderConfig)
-	if !ok {
-		resp.Diagnostics.AddError(
-			"Unexpected Data Source Configure Type",
-			fmt.Sprintf("Expected *ProviderConfig, got: %T", req.ProviderData),
-		)
-		return
-	}
-
-	d.client = client
-}
-
-func (d *edgeConfigurationsDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
-	var data edgeConfigurationsDataSourceModel
-	resp.Diagnostics.Append(req.Config.Get(ctx, &data)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	client := d.client.omniAPI
-	organizationID := data.OrganizationID.ValueString()
-	clusterID := data.ClusterID.ValueString()
-	edgeLocationID := data.EdgeLocationID.ValueString()
-
-	params := omni.EdgeConfigurationsAPIListEdgeConfigurationsParams{
-		EdgeLocationId: &edgeLocationID,
-	}
-
-	apiResp, err := client.EdgeConfigurationsAPIListEdgeConfigurationsWithResponse(ctx, organizationID, clusterID, &params)
-	if err != nil {
-		resp.Diagnostics.AddError("Failed to list edge configurations", err.Error())
-		return
-	}
-
-	if apiResp.StatusCode() != http.StatusOK {
-		resp.Diagnostics.AddError(
-			"Failed to list edge configurations",
-			fmt.Sprintf("unexpected status code: %d, body: %s", apiResp.StatusCode(), string(apiResp.Body)),
-		)
-		return
-	}
-
-	if apiResp.JSON200 == nil || apiResp.JSON200.Items == nil {
-		data.Configurations = []edgeConfigurationDataModel{}
-		resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
-		return
-	}
-
-	configurations := apiResp.JSON200.Items
-
-	// Apply optional filters
-	filterID := data.ID.ValueString()
-	filterName := data.Name.ValueString()
-
-	var filteredConfigs []omni.EdgeConfiguration
-	for _, config := range *configurations {
-		if filterID != "" && (config.Id == nil || *config.Id != filterID) {
-			continue
-		}
-		if filterName != "" && config.Name != filterName {
-			continue
-		}
-		filteredConfigs = append(filteredConfigs, config)
-	}
-
-	// Convert to model
-	data.Configurations = make([]edgeConfigurationDataModel, 0, len(filteredConfigs))
-	for _, config := range filteredConfigs {
-		configModel := edgeConfigurationDataModel{
-			ID:             types.StringPointerValue(config.Id),
-			Name:           types.StringValue(config.Name),
-			EdgeLocationID: types.StringPointerValue(config.EdgeLocationId),
-		}
-
-		if config.Default != nil {
-			configModel.Default = types.BoolValue(*config.Default)
-		}
-		if config.UserDataBase64 != nil {
-			configModel.UserDataBase64 = types.StringValue(*config.UserDataBase64)
-		}
-
-		// AWS configuration
-		if config.Aws != nil {
-			configModel.Aws = &awsConfigurationDataModel{
-				ImageID:         types.StringPointerValue(config.Aws.ImageId),
-				BootDiskSizeGiB: int32ToInt64(config.Aws.BootDiskSizeGib),
-				Tags:            stringMapToTF(config.Aws.Tags),
-			}
-		}
-
-		// GCP configuration
-		if config.Gcp != nil {
-			configModel.Gcp = &gcpConfigurationDataModel{
-				ImageID:         types.StringPointerValue(config.Gcp.ImageId),
-				BootDiskSizeGiB: int32ToInt64(config.Gcp.BootDiskSizeGib),
-				Labels:          stringMapToTF(config.Gcp.Labels),
-			}
-		}
-
-		// OCI configuration
-		if config.Oci != nil {
-			configModel.Oci = &ociConfigurationDataModel{
-				ImageID:         types.StringPointerValue(config.Oci.ImageId),
-				BootDiskSizeGiB: int32ToInt64(config.Oci.BootDiskSizeGib),
-				Tags:            stringMapToTF(config.Oci.Tags),
-			}
-		}
-
-		// Custom configuration
-		if config.Custom != nil {
-			configModel.Custom = customConfigToTF(config.Custom)
-		} else {
-			configModel.Custom = types.MapNull(types.StringType)
-		}
-
-		data.Configurations = append(data.Configurations, configModel)
-	}
-
-	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
-}
-
-// int32ToInt64 converts *int32 to types.Int64
-func int32ToInt64(v *int32) types.Int64 {
-	if v == nil {
-		return types.Int64Null()
-	}
-	return types.Int64Value(int64(*v))
-}
-
-// stringMapToTF converts map[string]string to types.Map
-func stringMapToTF(m *map[string]string) types.Map {
-	if m == nil {
-		return types.MapNull(types.StringType)
-	}
-
-	elements := make(map[string]attr.Value, len(*m))
-	for k, v := range *m {
-		elements[k] = types.StringValue(v)
-	}
-
-	result, diags := types.MapValue(types.StringType, elements)
-	if diags.HasError() {
-		return types.MapNull(types.StringType)
-	}
-	return result
-}
-
-// customConfigToTF converts map[string]interface{} to types.Map
-func customConfigToTF(m *omni.CustomCloudConfiguration) types.Map {
-	if m == nil || *m == nil {
-		return types.MapNull(types.StringType)
-	}
-
-	if len(*m) == 0 {
-		return types.MapValueMust(types.StringType, map[string]attr.Value{})
-	}
-
-	elements := make(map[string]attr.Value, len(*m))
-	for k, v := range *m {
-		if strVal, ok := v.(string); ok {
-			elements[k] = types.StringValue(strVal)
-		} else if v == nil {
-			elements[k] = types.StringNull()
-		} else {
-			elements[k] = types.StringValue(fmt.Sprintf("%v", v))
-		}
-	}
-
-	result, diags := types.MapValue(types.StringType, elements)
-	if diags.HasError() {
-		return types.MapNull(types.StringType)
-	}
-	return result
-}
