+++ castai/provider_framework.go
@@ -155,6 +155,6 @@ func (p *frameworkProvider) Resources(_ context.Context) []func() resource.Resou
 func (p *frameworkProvider) DataSources(_ context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
 		newOmniClusterDataSource,
-		newEdgeConfigurationsDataSource,
+		newEdgeConfigurationDataSource,
 	}
 }
